Use boost::optional<typeId> instead of broking abstraction
In certain case, universe::typeOf is not unable to compute the type of
an expression (in particular when it is not valid). Previously, in this
case, it returns -1, which is at the moment compatible with typeId
(size_t typedef). Now, it return a boost::none, and we keep the typeId
abstraction.