Implement a sync_remote_proxy, with timeout get
the remote_proxy_sync abstraction is build on top of remote_proxy, on
base on asynchronous operation. It is only to hide the "asynchronous
complexity" to the caller. get() can accept a timeout, in this case, the
remote_proxy_sync will returns boost::none if the timeout fires between
the end of operation