Revamp task representation
Task are now represented as recipe. A task can be constructed from its
parser representation, and after that valided in some ability context.
universe does not have anymore a add_task method.

Adapt the various call of task description in the code (it was
previously a simple PoD struct, it is now a more complex class).
Separate test_compiler_parser in two (the new part is in
test_compiler_task).