Fix a subtle bug in unifyM_candidate.
In some case, the set of possible expression related to a symbol is
empty, which means that the unification space is empty too.
It seems that we consider this case before, and the current code leads
to a nice UB (taking one element from an empty set and incrementing the
iterator).
To fix it, we throw a no_candidate exception if one candidate set is
empty and deal with it in try_to_infer.