Make remote_proxy_sync take a io_service in argument
Create a io_service in it is pretty bad, as it leads to have several
event loop in the same thread, which leads to ... nothing :D So pass the
io_service from the current thread :).

Adapt interface of model::execute to pass the io_service in parameter
everywhere too.

Adapt TU to pass this extra parameter