Rework execution stuff to work in an asynchronous way
We don't need to use a separate thread, with locking issue and this kind
of stuff. It will be more efficient because we dispatch the different
jobs in // (to get remote value).
Of course, the asynchronous model has some downside, in particular, the
need to really take care about "context".

Still need to update ability / logic_layer to use the new model