Fix an potentil bug in handle_arguments_computation
If we return an error as soon as we receive it, we will call the the cb,
and then the dtor for this function_execution before it get chance to
finish its computation. Future call to handle_arguments_computation for
this objet will lead to a crash (or another random error).

Now, store the first error, but wait for the terminaison of the
function before throwing it.