Fix a bug in a logger related to multiples connection request
When we have a lot of call of logger without explicit access to the
io_service, the logger will try to request multiples time the connect
procedure. It is wrong (in particular on non-Linux) because we will end
with multiples connection try on the same socket.
The solution is here to serializer logging information into a stack, and
only call the "write" / "connect-write" sequence is the queue was empty.