Let fact_db get sub_expression symbol from equal fact
Nowadays, equal fact are used by the logic_var db and will lead to
  - ok => nothing to do, nothing new
  - conflicting_facts => incoherency in the domain
  - permutation => we do some unification, exchange some symbols

so we have an empty list of sub_expression for the unification
predicate. To fix that, add to the ok and require_permutation structure
information about symbol we get in the unification process.

Update the unit test to reflect the new name of the structures
Update the fact unit-test to make sure we don't have wrong conversion
from tribool to bool.