[logic] Implement logic_var_db::deadapt
This method is the contrary of the method adapt. On the input of a
function call with logic var, generate the list of function call with
"outside" variable. As there is potentially a mapping 1 -> N between
logic_var and "normal var", there is potentially a lots of different
function_call for one base function_call