[compiler] Implement procedure at the recipe level
The syntax is something like :
  let f x y = fun { .... }

there are currently not possibility to return stuff from the function.
Morever, these functions are not real closure : they are expanded in the
recipe. Variable are automatically remapped to avoid name collision, but
the scope of the variable is still "recipe" one (even if nobody can't
really use it, as it has private name).