[compiler] Simplify expression_ast
Unary and binary expression was stored as expression parametrized by
their operator. It leads to a big variant to store expression_ast, and a
lot of template instanciation for the different operations on it(however
in certain case, it was pretty expressive).

Various template dispatching are replaced by simple switch / case.

It reduces both the time of compilation, the size of the code, and the
size of the generated code.