[model] Rework the implementation of abort
Change the information stored in an model::identifier, storing an index
in the current computation.
Add a method abort(size_t idx, cb) in abortable_computation, which abort
the entry idx of the current computation and call the callback when the
associated function finishes
Rework compute_abort to use this method. In this way, it gains the
possibility to abort any long-standing function.
Rework compute_ensure to return the correct mode::identifier.