[model] Fix abortable_computation::abort
If we are already at the end of the recipe (waiting for the terminaison
of the different continuous constraint) (wait_terminaison state), if we
receive an abort, the only thing to do is to do nothing.
If we recall terminaison() as done previously, the count of remaining
continuous constraint is incorrect, and leads to terminaison of the
recipe before correct cleaning.