[model] Fix a race in pause / resume handling in wait_computation
The invariant running ^ waiting was not respected, and so the code was
possibly wrong. If the method is paused while waiting state (but still
in running before) and then resumed in the same state, the resume part
will call compute, which will call async_wait when it has finished. Two
async_wait on the same time lead to an ECANCELED on the first one, and
so the whole method is aborted.